Moved around more files and deleted duplicates
