+++ src/BrowserRouter.jsx
@@ -1,20 +1,20 @@
 import {createBrowserRouter, RouterProvider} from "react-router-dom";
 import HomePage from './pages/HomePage/HomePage.jsx'
-import AboutPage from './pages/AboutPage.jsx'
+import AboutPage from './pages/AboutPage/AboutPage.jsx'
 import CarsPage from './pages/CarsPage/CarsPage.jsx'
 import LoginPage from './pages/LoginPage/LoginPage.jsx'
 import SignupPage from "./pages/SignupPage/SignupPage.jsx";
-import ReviewPage from "./pages/ReviewPage.jsx";
+import ReviewPage from "./pages/ReviewPage/ReviewPage.jsx";
 import MyPage from "./pages/MyPage/MyPage.jsx";
-import NotFoundPage from "./pages/NotFoundPage.jsx";
+import NotFoundPage from "./pages/NotFoundPage/NotFoundPage.jsx";
 import UserPage from "./pages/UserPage/UserPage.jsx";
-import ManageOwnedCarsPage from "./pages/ManageOwnedCarsPage.jsx";
-import SettingsPage from "./pages/SettingsPage.jsx";
+import ManageOwnedCarsPage from "./pages/ManageOwnedCarsPage/ManageOwnedCarsPage.jsx";
+import SettingsPage from "./pages/SettingsPage/SettingsPage.jsx";
 import UsersPage from "./pages/UsersPage/UsersPage.jsx";
-import OrdersPage from "./pages/OrdersPage.jsx";
+import OrdersPage from "./pages/OrdersPage/OrdersPage.jsx";
 import AddNewCar from "./components/AddNewCar.jsx";
 import App from "./App.jsx";
-import EditCarPage from "./pages/EditCarPage.jsx";
+import EditCarPage from "./pages/EditCarPage/EditCarPage.jsx";
 
 const router = createBrowserRouter([
     {
