+++ src/components/OwnedCarsList/OwnedCarsList.jsx
@@ -1,8 +1,8 @@
 
-import SearchableFieldTable from "./SearchableFieldTable/SearchableFieldTable.jsx";
+import SearchableFieldTable from "../SearchableFieldTable/SearchableFieldTable.jsx";
 import React from "react";
 import {Link} from "react-router-dom";
-import {fetchWithAuth} from "../static/js/auth.js";
+import {fetchWithAuth} from "../../static/js/auth.js";
 
 function EditCar({row}){
     return(
