+++ src/pages/ConflictPage/ConflictPage.jsx
