+++ src/pages/MyPage/MyPage.jsx
@@ -10,7 +10,7 @@ import {Outlet } from 'react-router-dom';
 import SettingsNavbar from "./SettingsNavbar.jsx";
 import useTitle from "../../components/useTitle.jsx";
 import CheckLogin from "../../components/CheckLogin.jsx";
-import AccessDeniedPage from "../AccessDeniedPage.jsx";
+import AccessDeniedPage from "../AccessDeniedPage/AccessDeniedPage.jsx";
 import Logout from "../../components/Logout.jsx";
 import "./MyPage.css";
 import React from "react";
